combine printFileSizes and removeFileNameHash to fileSizeReporter
