--- packages/react-dev-utils/removeFileNameHash.js
@@ -1,5 +1,0 @@
-module.exports = function removeFileNameHash(appBuild, fileName) {
-  return fileName
-    .replace(appBuild, '')
-    .replace(/\/?(.*)(\.\w+)(\.js|\.css)/, (match, p1, p2, p3) => p1 + p3);
-};
